requirements.txt
# Recommended Python 3.10–3.12 (Windows)
PySide6>=6.6
openai>=1.30.0
python-dotenv>=1.0.1
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pyttsx3==2.90
# Optional offline STT:
# vosk>=0.3.45
